Change promises for async await on server.js

+++ server/server.js
@@ -14,70 +14,75 @@ const port = process.env.PORT || 3000;
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req, res) => {
-    var todo = new Todo({
-        text: req.body.text,
-        _creator: req.user._id
-    });
+app.post('/todos', authenticate, async (req, res) => {
+    try {
+        const todo = new Todo({
+            text: req.body.text,
+            _creator: req.user._id
+        });
 
-    todo.save().then((doc) => {
+        const doc = await todo.save()
         res.send(doc);
-    }, (e) => {
-        res.status(400).send(e);
-    })
+
+    } catch (error) {
+        res.status(400).send(error);
+    }
 });
 
-app.get('/todos', authenticate, (req, res) => {
-    Todo.find({
-        _creator: req.user._id
-    }).then((todos) => {
+app.get('/todos', authenticate, async (req, res) => {
+
+    try {
+        const todos = await Todo.find({
+            _creator: req.user._id
+        });
         res.send({ todos });
-    }, (e) => {
-        res.status(400).send(e);
-    })
+    } catch (error) {
+        res.status(400).send(error);
+    }
 });
 
-app.get('/todos/:id', authenticate, (req, res) => {
-    var id = req.params.id;
+app.get('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id;
 
     if (!ObjectID.isValid(id)) {
         res.status(404).send();
     }
 
-    Todo.findOne({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
+    try {
+        const todo = await Todo.findOne({
+            _id: id,
+            _creator: req.user._id
+        });
+
         if (!todo) {
             res.status(404).send();
         }
         res.send({ todo });
-
-    }).catch((e) => res.status(400).send());
+    } catch (error) {
+        res.status(400).send()
+    }
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    var id = req.params.id;
+app.delete('/todos/:id', authenticate, async (req, res) => {
 
-    if (!ObjectID.isValid(id)) {
-        res.status(404).send();
-    }
+    try {
+        const id = req.params.id;
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
+        if (!ObjectID.isValid(id)) {
+            res.status(404).send();
+        }
+        const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
         if (!todo) {
             res.status(404).send();
         }
         res.send({ todo });
-
-    }).catch((e) => {
+    } catch (error) {
         res.status(400).send();
-    })
+    }
+
 });
 
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
     var id = req.params.id
     var body = _.pick(req.body, ['text', 'completed']);
 
@@ -91,31 +96,34 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         body.completed = false;
         body.completedAt = null;
     }
+   
 
-    Todo.findOneAndUpdate({
-        _id: id,
-        _creator: req.user._id
-    }, { $set: body }, { new: true }).then((todo) => {
+    try {
+        const todo = await Todo.findOneAndUpdate({
+            _id: id,
+            _creator: req.user._id
+        }, { $set: body }, { new: true });
         if (!todo) {
             res.status(404).send();
         }
+
         res.send({ todo });
-    }).catch((e) => {
+    } catch (error) {
         res.status(400).send();
-    })
+    }
+    
 })
 
-app.post('/users', (req, res) => {
-    var body = _.pick(req.body, ['email', 'password']);
-    var user = new User(body);
-
-    user.save().then(() => {
-        return user.generateAuthToken();
-    }).then((token) => {
+app.post('/users', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = new User(body);
+        await user.save();
+        const token = await user.generateAuthToken();
         res.header('x-auth', token).send(user);
-    }).catch((e) => {
-        res.status(400).send(e);
-    })
+    } catch (error) {
+        res.status(400).send(error);
+    }
 })
 
 app.get('/users/me', authenticate, (req, res) => {
@@ -123,24 +131,27 @@ app.get('/users/me', authenticate, (req, res) => {
 })
 
 
-app.post('/users/login', (req, res) => {
-    var body = _.pick(req.body, ['email', 'password']);
+app.post('/users/login', async (req, res) => {
 
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });
-    }).catch((e) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken()
+        res.header('x-auth', token).send(user);
+    } catch (error) {
         res.status(400).send();
-    })
+    }
+
 })
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token);
         res.send();
-    }, () => {
+    } catch (error) {
         res.status(400).send();
-    })
+    }
+
 });
 
 app.listen(port, () => {
